+++ gui-progress/src/filecopier.cpp
@@ -25,8 +25,7 @@ void FileCopier::copyFile(const QString& source,const QString& destination)
     qint64 size = sourceFile.size();
     QByteArray buffer(1024*1024,'\0');
     qint64 count,copied = 0;
-    // После каждого блока проверяем, не отменена ли операция.
-    // Явного volatile на cancelled не требуется, он содержится в его в реализации.
+    // Manually check whether operation was cancelled after every block.
     while((count = sourceFile.read(buffer.data(),buffer.size()))>0&&!cancelled_){
         if(destinationFile.write(buffer.data(),count)!=count){
             emit done(tr("Failed to write to destination file: %1").arg(
@@ -34,10 +33,10 @@ void FileCopier::copyFile(const QString& source,const QString& destination)
             return;
         }
         copied += count;
-        int newProgress = 100*copied/size;
+        int newProgress = int(100*copied/size);
         if(progress!=newProgress){
-            // Посылаем сигнал обновления прогресса только когда меняется
-            // целое число процентов, чтобы не обновлять интерфейс слишком часто.
+            // We send an update only for whoole percent changes
+            // to avoid spending too much time sending UI updates.
             progress = newProgress;
             emit progressChanged(progress);
         }
